internal refactoring
move out the MetadataHandler in a separate module
add decorator to measure method time execution
