+++ source/messages.py
@@ -47,11 +47,12 @@ MSG = {'IntError': 'Server internal error occurred. Reason: {}',
        'BucketsizeChange': 'Based on requested downsample value: {} the bucketsize will be set: {}',
        'BucketsizeToPeriod': 'Bucketsize will be set to sensors period: {}',
        'ReceivedQuery': 'Received query request for query:{}, start:{}, end:{}',
-       'RunQuery': 'Execute zimon query: {}',
        'AttrNotValid': 'Invalid attribute:{}',
        'AllowedAttrValues': 'For attribute {} applicable values:{}',
        'ReceivAttrValues': 'Received {}:{}',
-       'TimerInfo': 'Processing {} took {} seconds',
+       'StartMethod': 'Starting method: {} with args: {}.',
+       'RunMethod': 'Executed method: {} with args: {}.',
+       'TimerInfo': 'Processing {} took {:.3f} seconds.',
        'Query2port': 'For better bridge performance multithreaded port {} will be used',
        'CollectorConnInfo': 'Connection to the collector server established successfully',
        'BridgeVersionInfo': ' *** IBM Spectrum Scale bridge for Grafana - Version: {} ***',
